smart skipFrames, fixed memory pool

+++ include/decord/video_interface.h
@@ -59,7 +59,7 @@ class VideoReaderInterface {
     /*! \brief retrieve keyframe indices */
     virtual runtime::NDArray GetKeyIndices() = 0;
     /*! \brief skip certain frames without decoding */
-    virtual void SkipFrames(int num = 1) = 0;
+    virtual void SkipFrames(int64_t num = 1) = 0;
     /*! \brief destructor */
     virtual ~VideoReaderInterface() = default;
