+++ src/video/video_interface.cc
@@ -67,7 +67,7 @@ DECORD_REGISTER_GLOBAL("video_reader._CAPI_VideoSeek")
 DECORD_REGISTER_GLOBAL("video_reader._CAPI_VideoSkipFrames")
 .set_body([] (DECORDArgs args, DECORDRetValue* rv) {
     VideoReaderInterfaceHandle handle = args[0];
-    int num = args[1];
+    int64_t num = args[1];
     static_cast<VideoReaderInterface*>(handle)->SkipFrames(num);
   });
 }  // namespace runtime
